feat: implement delta time into the engine class and utilize it in the camera class

+++ include/camera.h
@@ -10,19 +10,19 @@
 class Camera
 {
 public:
-    Camera(glm::vec3 position, int width, int height);
+    Camera(glm::vec3 position, unsigned int width, unsigned int height);
 
     void Matrix(float fov, float near_plane, float far_plane, unsigned int shader_program_id, const char* uniform) const;
 
-    void HandleInput(GLFWwindow* window, float delta_time);
+    void HandleInput(GLFWwindow* window, double delta_time);
 
 private:
     glm::vec3 position_;
     glm::vec3 orientation_;
     glm::vec3 up_;
 
-    int width_;
-    int height_;
+    unsigned int width_;
+    unsigned int height_;
 
     float speed_;
     float sensitivity_;
